% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_cloud.R
\name{browse_cloud}
\alias{browse_cloud}
\title{Open learning materials in Posit Cloud}
\usage{
browse_cloud(module, open = interactive())
}
\arguments{
\item{module}{The name of the module}

\item{open}{Logical. Open the URL in the browser? Default is \code{TRUE}. If
\code{FALSE}, \code{browse_cloud()} just provides the URL for you to open yourself.}
}
\description{
\code{browse_cloud()} will open an Posit Cloud version of a given learning
module in your browser. The cloud version will have the packages and
materials pre-installed.
}
