% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_module.R
\name{use_module}
\alias{use_module}
\title{Install learning materials}
\usage{
use_module(module, destdir = NULL)
}
\arguments{
\item{module}{The name of the module}

\item{destdir}{The path on your computer where you would like the learning
materials installed. If \code{NULL}, defaults to your Desktop or some other
conspicuous place.}
}
\description{
\code{use_module()} will install the materials for a given learning module. Then,
it will open a new RStudio Project containing the files you'll need for that
module.
}
